Si570 Clock Controller
Implementation of a Si570 initialization controller.

The Si570 has a very complex initialization routine so the logic to set
it up to an arbitary frequency isn't straight forward.

This core can easily be re-purposed to allow for dynamic clock setup as
well. Currently it is static to allow whatever optimizations that can be
done to be done.