fejkon_pcie_avalon: support slow reads
Add proper support for waitrequest and change the pcie-hip testbench to
smash more on the PCIe core to weed out issues when there is pressure on
the Avalon-MM interface.

Right now it seems we are capable of doing up to 32 write/reads queued,
probably more in real life. We are very unlikely to actually do that
bursty read/writes.